+++ src/main/java/com/bestseller/coffeestore/admin/model/Topping.java
@@ -1,17 +1,11 @@
 package com.bestseller.coffeestore.admin.model;
 
-import java.math.BigInteger;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+
+import javax.persistence.*;
 
 @Getter
 @Setter
@@ -26,5 +20,5 @@ public class Topping {
 
 	String name;
 
-	BigInteger price;
+	Double price;
 }
